pyproject.toml
[project]
name = "Conversational_AI_Agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
   "black>=23.12.0",
   "boto3>=1.39.4",
   "botocore>=1.39.4",
   "faiss-cpu>=1.11.0",
   "fastapi>=0.109.0",
   "httpx>=0.28.1",
   "isort>=5.13.0",
   "joblib>=1.5.1",
   "langchain>=0.1.0",
   "langchain-community>=0.3.26",
   "langgraph>=0.0.20",
   "langsmith>=0.0.69",
   "mypy>=1.8.0",
   "openai>=1.90.0",
   "peft>=0.15.2",
   "pydantic>=2.5.0",
   "pydantic-settings>=2.9.1",
   "pymupdf>=1.26.1",
   "pytest>=7.4.0",
   "python-dotenv>=1.0.0",
   "sentence-transformers>=4.1.0",
   "streamlit>=1.31.0",
   "torch>=2.1.0",
   "transformers>=4.36.0",
   "typing-extensions>=4.14.0",
   "uvicorn>=0.27.0",
]

[dependency-groups]
dev = [
    "black>=23.12.0",
    "boto3>=1.39.4",
    "botocore>=1.39.4",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.109.0",
    "httpx>=0.28.1",
    "isort>=5.13.0",
    "joblib>=1.5.1",
    "langchain>=0.1.0",
    "langchain-community>=0.3.26",
    "langgraph>=0.0.20",
    "langsmith>=0.0.69",
    "mypy>=1.8.0",
    "openai>=1.90.0",
    "peft>=0.15.2",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.9.1",
    "pymupdf>=1.26.1",
    "pytest>=7.4.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=4.1.0",
    "streamlit>=1.31.0",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "typing-extensions>=4.14.0",
    "uvicorn>=0.27.0",
]

[tool.uv.sources]
conversational-ai-agent = { workspace = true }
